requirements.txt
web3==6.11.1
pandas==2.1.3
openai==1.3.7
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
scikit-learn==1.3.2 